wip: text editor unit tests
