+++ packages/web-app-pdf-viewer/src/App.vue
@@ -1,88 +1,29 @@
 <template>
-  <main>
-    <app-top-bar v-if="!loading && !loadingError" :resource="resource" @close="closeApp" />
-    <loading-screen v-if="loading" />
-    <error-screen v-else-if="loadingError" />
-    <div v-else class="oc-height-1-1">
-      <object class="pdf-viewer oc-height-1-1 oc-width-1-1" :data="url" :type="objectType" />
-    </div>
-  </main>
+  <object class="pdf-viewer oc-width-1-1 oc-height-1-1" :data="url" type="application/pdf" />
 </template>
+
 <script lang="ts">
-import { computed, defineComponent } from 'vue'
-import { useAppDefaults } from 'web-pkg/src/composables'
-import AppTopBar from 'web-pkg/src/components/AppTopBar.vue'
-import ErrorScreen from './components/ErrorScreen.vue'
-import LoadingScreen from './components/LoadingScreen.vue'
+import { defineComponent } from 'vue'
 
 export default defineComponent({
-  name: 'PDFViewer',
-  components: {
-    AppTopBar,
-    ErrorScreen,
-    LoadingScreen
+  props: {
+    url: {
+      type: String,
+      required: true
+    }
   },
   setup() {
-    const isSafari = () =>
+    const isSafari =
       navigator.userAgent?.includes('Safari') && !navigator.userAgent?.includes('Chrome')
 
-    const objectType = computed(() => {
-      // object type must not be 'application/pdf' for Safari due to a bug
-      return isSafari() ? undefined : 'application/pdf'
-    })
+    const objectType = isSafari ? undefined : 'application/pdf'
     return {
-      ...useAppDefaults({
-        applicationId: 'pdf-viewer'
-      }),
       objectType
-    }
-  },
-  data: () => ({
-    loading: true,
-    loadingError: false,
-    url: '',
-    resource: null
-  }),
-  watch: {
-    currentFileContext: {
-      handler: function () {
-        this.unloadPdf()
-        this.loadPdf(this.currentFileContext)
-      },
-      immediate: true
-    }
-  },
-  beforeUnmount() {
-    this.unloadPdf()
-  },
-  methods: {
-    async loadPdf(fileContext) {
-      try {
-        this.loading = true
-        const resource = await this.getFileInfo(fileContext)
-
-        if (resource.mimeType !== 'application/pdf') {
-          return
-        }
-
-        this.resource = resource
-        this.replaceInvalidFileRoute(this.currentFileContext, this.resource)
-        this.url = await this.getUrlForResource(fileContext.space, this.resource, {
-          disposition: 'inline'
-        })
-      } catch (e) {
-        this.loadingError = true
-        console.error('Error fetching pdf', e)
-      } finally {
-        this.loading = false
-      }
-    },
-    unloadPdf() {
-      this.revokeUrl(this.url)
     }
   }
 })
 </script>
+
 <style scoped>
 .pdf-viewer {
   border: none;
