+++ packages/web-app-pdf-viewer/src/index.ts
@@ -1,6 +1,6 @@
 import translations from '../l10n/translations.json'
 import { AppWrapperRoute } from 'web-pkg/src/components/AppTemplates/AppWrapperRoute'
-import PdfViewer from './PdfViewer.vue'
+import PdfViewer from './App.vue'
 
 // just a dummy function to trick gettext tools
 function $gettext(msg) {
